Added php files
The following php files are responsible for establishing a connection with a local server, verifying login credentials, and accepting data for new users respectively, wamp64 server was used to run the server holding the database. So far the register.php script only accepts a name, username, and password. More fields will be permitted in the future.